refactor: 예외 처리 설정 변경
- 계층 정보 제외
