+++ src/main/kotlin/com/deepromeet/atcha/common/exception/CustomException.kt
@@ -1,0 +1,12 @@
+package com.deepromeet.atcha.common.exception
+
+abstract class CustomException(
+    val errorType: BaseErrorType
+) : RuntimeException() {
+
+    val status: Int
+        get() = errorType.errorReason.status
+
+    override val message: String
+        get() = errorType.errorReason.message
+}
