--- src/main/kotlin/com/deepromeet/seulseul/common/exception/CustomException.kt
@@ -1,14 +1,0 @@
-package com.deepromeet.seulseul.common.exception
-
-abstract class CustomException(
-    val errorType: BaseErrorType,
-    private val sourceLayer: String,
-) : RuntimeException() {
-
-    val status: Int
-        get() = errorType.errorReason.status
-
-    override val message: String
-        get() =
-                sourceLayer.let { "$it - ${errorType.errorReason.message}" }
-}
