--- src/main/kotlin/com/deepromeet/seulseul/common/exception/DomainException.kt
@@ -1,5 +1,0 @@
-package com.deepromeet.seulseul.common.exception
-
-open class DomainException(errorCode: BaseErrorType) : CustomException(errorCode, "도메인 계층 예외") {
-
-}
