--- src/main/kotlin/com/deepromeet/seulseul/common/exception/InfraException.kt
@@ -1,4 +1,0 @@
-package com.deepromeet.seulseul.common.exception
-
-open class InfraException(errorCode: BaseErrorType) :
-    CustomException(errorCode, "인프라 계층 예외")
