--- src/main/kotlin/com/deepromeet/seulseul/common/exception/WebException.kt
@@ -1,4 +1,0 @@
-package com.deepromeet.seulseul.common.exception
-
-open class WebException(errorCode: BaseErrorType) :
-        CustomException(errorCode, "웹 계층 예외")
